Update homebrew.md
Blitzer shall be renamed to "Firestorm Weapon," with the Heavy and Two-handed properties as requirements. The final paragraph which reads "When you roll a 20 on an attack roll made with this weapon, you can attempt to make a Firestorm strike on a creature when you hit. The creature must succeed a DC 15 Dexterity saving throw or take an additional 8d4 fire damage" shall now read the following:

"When you attack a creature with this magic weapon and roll a 20 on the attack roll, that creature must make a DC 15 Dexterity saving throw. The creature takes 6d6 fire damage on a failed save, or half as much damage on a successful one."

Channeler's Trident shall now be known as "Crusader's Armanent" and may be any type of weapon.